Rename std/crypto to std/hash
