+++ std/hash/sha1.ts
